+++ src/core/shims/global.js
@@ -13,10 +13,13 @@ const _global =
 	typeof window === 'object' && isGlobal(window) && window ||
 	typeof global === 'object' && isGlobal(global) && global ||
 	typeof self === 'object' && isGlobal(self) && self ||
+
 	(function getGlobalUnstrict() {
 		return this;
 	}()) ||
+
 	this ||
+
 	// eslint-disable-next-line no-new-func
 	new Function('', 'return this')();
 
@@ -24,7 +27,7 @@ module.exports = _global;
 
 /**
  * Checks if the provided value is a global object by confirming the presence of Math,
- * known to exist in any global JavaScript environment.
+ * known to exist in any global JS environment
  *
  * @param obj
  * @returns {boolean}